+++ samples/index.php
@@ -1,5 +1,7 @@
 <?php
 include_once 'Sample_Header.php';
+
+use PhpOffice\PhpWord\Settings;
 $requirements = array(
     'php'   => array('PHP 5.3.0', version_compare(phpversion(), '5.3.0', '>=')),
     'xml'   => array('PHP extension XML', extension_loaded('xml')),